Resolve consistency for editor source filenames.
