+++ src/engine/editor/ActorTreeWindow.h
@@ -17,31 +17,17 @@
 
 #pragma once
 
-#include "base_window.h"
+#include "BaseWindow.h"
 
-namespace ohw {
-	class Map;
-	class MapConfigEditor : public BaseWindow {
-	public:
-		MapConfigEditor();
-		~MapConfigEditor() override;
+class ActorTreeWindow : public BaseWindow {
+public:
+	ActorTreeWindow();
+	~ActorTreeWindow() override;
 
-		void Display() override;
-		void SaveManifest( const std::string &path );
-		void RestoreManifest();
+	void Display() override;
+	void DisplayActorProperties( Actor* actor );
 
-	protected:
-	private:
-		MapManifest backup_;
-		MapManifest *manifest_;
-		Map *map_{ nullptr };
-
-		char name_buffer[32]{ '\0' };
-		char author_buffer[32]{ '\0' };
-		char filename_buffer[32]{ '\0' };
-
-		void DisplayWeather();
-		void DisplayTemperature();
-		void DisplayTime();
-	};
-}
+protected:
+private:
+	bool showReadOnly{ false };
+};
