+++ src/engine/editor/MapConfigEditor.h
@@ -17,15 +17,31 @@
 
 #pragma once
 
-#include "base_window.h"
+#include "BaseWindow.h"
 
-class NewMapWindow : public BaseWindow {
- public:
+namespace ohw {
+	class Map;
+	class MapConfigEditor : public BaseWindow {
+	public:
+		MapConfigEditor();
+		~MapConfigEditor() override;
 
-  void Display() override;
+		void Display() override;
+		void SaveManifest( const std::string &path );
+		void RestoreManifest();
 
- protected:
- private:
-  char name_buffer_[32]{'\0'};
-  char author_buffer_[32]{'\0'};
-};
+	protected:
+	private:
+		MapManifest backup_;
+		MapManifest *manifest_;
+		Map *map_{ nullptr };
+
+		char name_buffer[32]{ '\0' };
+		char author_buffer[32]{ '\0' };
+		char filename_buffer[32]{ '\0' };
+
+		void DisplayWeather();
+		void DisplayTemperature();
+		void DisplayTime();
+	};
+}
