+++ src/engine/editor/TerrainImportWindow.cpp
@@ -18,25 +18,41 @@
 #include <imgui.h>
 
 #include "../engine.h"
+#include "../Map.h"
+#include "../Terrain.h"
 
-#include "window_new_map.h"
+#include "TerrainImportWindow.h"
 
 using namespace ohw;
 
-void NewMapWindow::Display() {
-	ImGui::SetNextWindowSize( ImVec2( 256, 128 ), ImGuiCond_Once );
-	ImGui::Begin( "New Map", &status_, ED_DEFAULT_WINDOW_FLAGS );
-	ImGui::InputText( "Name", name_buffer_, sizeof( name_buffer_ ) );
-	//ImGui::InputText("Author", author_buffer_, sizeof(author_buffer_));
-	ImGui::Separator();
-	if ( ImGui::Button( "Create" ) ) {
-		Engine::Game()->CreateManifest( name_buffer_ );
-		Engine::Game()->LoadMap( name_buffer_ );
-		SetStatus( false );
+TerrainImportWindow::TerrainImportWindow() = default;
+TerrainImportWindow::~TerrainImportWindow() = default;
+
+void TerrainImportWindow::Display() {
+	ImGui::SetNextWindowSize( ImVec2( 310, 128 ), ImGuiCond_Once );
+	Begin( "Import Heightmap", ED_DEFAULT_WINDOW_FLAGS );
+	ImGui::InputText( "Path", path_buffer, sizeof( path_buffer ) );
+	ImGui::InputInt( "Multiplier", &multiplier_ );
+	if ( ImGui::Button( "Import" ) ) {
+		ImportTerrain();
 	}
 	ImGui::SameLine();
 	if ( ImGui::Button( "Cancel" ) ) {
 		SetStatus( false );
 	}
 	ImGui::End();
+}
+
+void TerrainImportWindow::ImportTerrain() {
+	Map *map = Engine::Game()->GetCurrentMap();
+	if ( map == nullptr ) {
+		return;
+	}
+
+	Terrain *terrain = map->GetTerrain();
+	if ( terrain == nullptr ) {
+		return;
+	}
+
+	terrain->LoadHeightmap( path_buffer, multiplier_ );
 }
